Implement STORAGE functionality
This patch adds a Storage object to the P2PHandler in order to answer
incoming STORAGE GET and STORAGE PUT requests.

The current implementation is inefficient since the whole Storage object
is locked for the full time the request is handled. Instead, we want
parallel IO operations and only lock the single data object while
reading it.