Add support for config file
This commit allows to pass a -c or --config parameter which contains the
path to a config ini file as required from the specification.

The relevant `dht` section from the config file is read and the relevant
information stored in a Config struct which is then passed to the main
function of our dht program.